+++ include/swift/SILOptimizer/PassManager/Transforms.h
@@ -81,7 +81,6 @@ namespace swift {
     /// Get the transform's name as a C++ identifier.
     llvm::StringRef getID() { return PassKindID(getPassKind()); }
 
-  protected:
     /// Searches for an analysis of type T in the list of registered
     /// analysis. If the analysis is not found, the program terminates.
     template<typename T>
