+++ test/SILOptimizer/simplify_cfg_ossa_simplify_branch.sil
@@ -1,4 +1,4 @@
-// RUN: %target-sil-opt -unit-test-runner %s 2>&1 | %FileCheck %s
+// RUN: %target-sil-opt -test-runner %s 2>&1 | %FileCheck %s
 
 enum FakeOptional<T> {
   case some(T)
