+++ test/SILOptimizer/simplify_switch_enum_objc_ossa.sil
@@ -1,4 +1,4 @@
-// RUN: %target-sil-opt -unit-test-runner %s 2>&1 | %FileCheck %s
+// RUN: %target-sil-opt -test-runner %s 2>&1 | %FileCheck %s
 
 // REQUIRES: objc_interop
 import Swift
